fix: report package version in server info
