+++ src/index.ts
@@ -19,28 +19,11 @@
  *                         wait_for_all, stop_agent, kill, interact
  */
 
-import { readFileSync } from "node:fs";
-import { dirname, join } from "node:path";
-import { fileURLToPath } from "node:url";
 import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js";
 import { createServer } from "./server.js";
 import { createCmuxClient } from "./cmux-client-factory.js";
 import { renderDoctorJson, renderDoctorText, runDoctor } from "./doctor.js";
-
-function readVersion(): string {
-  // package.json sits one level above the compiled entrypoint (dist/index.js
-  // → ../package.json, and likewise libexec/dist/index.js → libexec/package.json
-  // for a brew install). Best-effort; never throw from a --version probe.
-  try {
-    const here = dirname(fileURLToPath(import.meta.url));
-    const pkg = JSON.parse(
-      readFileSync(join(here, "..", "package.json"), "utf-8"),
-    ) as { version?: string };
-    return pkg.version ?? "unknown";
-  } catch {
-    return "unknown";
-  }
-}
+import { readVersion } from "./version.js";
 
 const HELP_TEXT = `cmuxlayer — Terminal multiplexer MCP server for AI agent workspace orchestration.
